refactor raw stats converter and fix style errors
